# api-spec.yaml
openapi: 3.0.3
info:
  title: StreamVault API
  description: |
    Professional video hosting platform API. Authenticate with a JWT Bearer token or API key (`sv_live_xxx`).
    
    All timestamps are **UNIX epoch integers** (seconds).
  version: 1.0.0
  contact:
    name: StreamVault Support
    url: https://streamvault.link
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.streamvault.link
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      in: header
      name: X-Workspace-Id
      required: true
      schema:
        type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/folders:
    get:
      tags: [Folders]
      summary: List all folders in workspace
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Folder list
    post:
      tags: [Folders]
      summary: Create folder
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                parent_id: { type: string, nullable: true }
      responses:
        '201':
          description: Folder created
  /api/folders/{id}:
    patch:
      tags: [Folders]
      summary: Rename folder
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200':
          description: Renamed
    delete:
      tags: [Folders]
      summary: Delete folder (videos become unfoldered)
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /api/playlists:
    get:
      tags: [Playlists]
      summary: List playlists in workspace
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Playlist list with video counts
    post:
      tags: [Playlists]
      summary: Create playlist
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                visibility: { type: string, enum: [public, private] }
      responses:
        '201':
          description: Playlist created
  /api/playlists/{id}/videos:
    get:
      tags: [Playlists]
      summary: Get videos in playlist (ordered by position)
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Video list with position field
    post:
      tags: [Playlists]
      summary: Add video to playlist
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [video_id]
              properties:
                video_id: { type: string }
      responses:
        '201':
          description: Video added
        '409':
          description: Video already in playlist
...